Refactoring part 2/2 B)
