+++ src/appstate.rs
@@ -1,3 +1,5 @@
+// Froggi appstate
+
 use std::sync::Arc;
 use tokio::sync::Mutex;
