+++ src/routing/basic.rs
@@ -1,3 +1,5 @@
+// Froggi routing (basic)
+
 use axum::{
     body::Body,
     http::{header::LOCATION, HeaderValue, StatusCode},
