+++ src/routing/team.rs
@@ -1,3 +1,5 @@
+// Froggi routing (teams)
+
 use axum::{
     extract::{Path, State},
     http::StatusCode,
