+++ src/routing/updating.rs
@@ -1,13 +1,11 @@
-use anyhow::{anyhow, Result};
+// Froggi routing (updating)
+
 use axum::{
     http::StatusCode,
     response::{Html, IntoResponse, Response},
 };
 
-use crate::{appstate::global::*, printlg};
-
-const REMOTE_CARGO_TOML_URL: &'static str =
-    "https://raw.githubusercontent.com/AllLiver/Froggi/refs/heads/main/Cargo.toml";
+use crate::{appstate::global::*, printlg, update_checker, REMOTE_CARGO_TOML_URL};
 
 pub async fn update_handler() -> impl IntoResponse {
     printlg!("Checking if an update is availible...");
@@ -44,61 +42,6 @@ pub async fn update_handler() -> impl IntoResponse {
                 REMOTE_CARGO_TOML_URL
             ))
             .unwrap();
-    }
-}
-
-pub async fn update_checker() -> Result<(bool, String)> {
-    let result = reqwest::get(REMOTE_CARGO_TOML_URL).await;
-
-    if let Ok(response) = result {
-        let remote_version_str_raw = response.text().await.expect("Failed to get response text");
-        let remote_version_str = remote_version_str_raw
-            .split("\n")
-            .collect::<Vec<&str>>()
-            .iter()
-            .find(|x| x.starts_with("version = "))
-            .expect("Failed to get remote version")
-            .trim_start_matches("version = \"")
-            .trim_end_matches("\"");
-
-        let local_version_str = env!("CARGO_PKG_VERSION");
-
-        let remote_version: Vec<u8> = remote_version_str
-            .split(".")
-            .map(|x| x.parse::<u8>().expect("Failed to parse remote version"))
-            .collect();
-        let local_version: Vec<u8> = local_version_str
-            .split(".")
-            .map(|x| x.parse::<u8>().expect("Failed to parse remote version"))
-            .collect();
-
-        let mut out_of_date = false;
-
-        for i in 0..local_version.len() {
-            if remote_version[i] > local_version[i] {
-                out_of_date = true;
-                break;
-            } else if remote_version[i] < local_version[i] {
-                break;
-            }
-        }
-
-        return Ok((out_of_date, String::from(remote_version_str)));
-    } else if let Err(e) = result {
-        return Err(anyhow!("{}", e));
-    } else {
-        return Err(anyhow!("Some unexpected error"));
-    }
-}
-
-pub async fn auto_update_checker() {
-    loop {
-        if let Ok(r) = update_checker().await {
-            *OUT_OF_DATE.lock().await = r.0;
-            *REMOTE_VERSION.lock().await = r.1;
-        }
-
-        tokio::time::sleep(std::time::Duration::from_secs(60 * 10)).await;
     }
 }
