+++ src/routing/visibility.rs
@@ -1,3 +1,5 @@
+// Froggi routing (visibility)
+
 use axum::{
     extract::{Path, State},
     http::StatusCode,
