+++ src/routing/websockets.rs
@@ -1,3 +1,5 @@
+// Froggi routing (websockets)
+
 use axum::{
     extract::{ws::Message, State, WebSocketUpgrade},
     response::IntoResponse,
